refactor: 拡張性のため、予報系のstoreとinterfaceはグループ化
