+++ app/interfaces/weatherForecast/MeteorologicalObservatoryInterface.ts
