+++ app/interfaces/weatherForecast/WeatherForecastInterface.ts
