+++ app/interfaces/weatherForecast/WeatherForecastQueryParamsInterface.ts
